CohortSymmetry: Sequence Symmetry Analysis Using the Observational Medical Outcomes Partnership Common Data Model

Calculating crude sequence ratio, adjusted sequence ratio and confidence intervals using data mapped to the Observational Medical Outcomes Partnership Common Data Model.

See also

Author

Maintainer: Xihang Chen xihang.chen@ndorms.ox.ac.uk (ORCID)

Authors:

Tyman Stanford tyman.stanford@unisa.edu.au (ORCID)

Berta Raventós braventos@idiapjgol.info (ORCID)

Nicole Pratt Nicole.Pratt@unisa.edu.au (ORCID)

Ed Burn ed.burn@ndorms.ox.ac.uk (ORCID)

Marti Català marti.catalasabate@ndorms.ox.ac.uk (ORCID)

Danielle Newby danielle.newby@ndorms.ox.ac.uk (ORCID)

Núria Mercadé-Besora nuria.mercadebesora@ndorms.ox.ac.uk (ORCID)

Mike Du mike.du@ndorms.ox.ac.uk (ORCID)

Yuchen Guo yuchen.guo@ndorms.ox.ac.uk (ORCID)

Kim Lopez kim.lopez@spc.ox.ac.uk (ORCID)

Marta Alcalde-Herraiz marta.alcaldeherraiz@ndorms.ox.ac.uk (ORCID)