A formatted visualization of sequence_symmetry objects.

It provides a formatted table with the contents of the getSequenceRatios output.

Usage

formatSequenceSymmetry(
  result,
  type = "gt",
  estimateNameFormat = c(`N (%)` = "<count> (<percentage> %)", `SR (CI)` =
    "<point_estimate> (<lower_CI> - <upper_CI>)"),
  style = "default",
  crude = TRUE,
  adjusted = TRUE,
  studyPopulation = TRUE,
  indexName = TRUE,
  markerName = TRUE,
  cdmName = TRUE,
  .options = NULL
)

Arguments

result

A sequence_symmetry object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

estimateNameFormat

.

style

Named list that specifies how to style the different parts of a gt table or flextable. See visOmopResults package for more information on how to define a style. Alternatively, use "default" to get visOmopResults style, or NULL for gt/flextable default styling.

crude

whether to report crude results.

adjusted

whether to report adjusted results.

studyPopulation

whether to report the study population.

indexName

whether to report index names.

markerName

whether to report marker names

cdmName

whether to report database names.

.options

named list with additional formatting options. CohortSymmetry::formatSequenceSymmetryOptions() shows allowed arguments and their default values.

Value

A formatted version of the sequence_symmetry object.