A tidy visualization of sequence_symmetry objects.

It provides a tidy dataframe with the contents of the summariseSequenceRatio output.

Usage

tidySequenceSymmetry(result)

Arguments

result

A sequence_symmetry object.

Value

A tibble with a tidy version of the sequence_symmetry object.

Examples

# \donttest{
library(CohortSymmetry)
cdm <- mockCohortSymmetry()
cdm <- CohortSymmetry::generateSequenceCohortSet(cdm = cdm,
                                                 name = "joined_cohorts",
                                                 indexTable = "cohort_1",
                                                 markerTable = "cohort_2",
                                                 combinationWindow = c(0, Inf))
res <- CohortSymmetry::summariseSequenceRatio(
cdm = cdm,
sequenceCohortSet = "joined_cohorts")
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events
#> Warning: NSR numerator is 0, which results in a NSR = 0, proceed with caution
#> Warning: NSR denominator is 0, suggesting no Marker Drug -> Index Drug or Index Drug -> Marker Drug events

tidy_result <- CohortSymmetry::tidySequenceSymmetry(res)

tidy_result
#> # A tibble: 8 × 20
#>   result_id cdm_name index_cohort_name marker_cohort_name days_prior_observation
#>   <chr>     <chr>    <chr>             <chr>              <chr>                 
#> 1 2         mock da… cohort_1          cohort_2           0                     
#> 2 3         mock da… cohort_1          cohort_3           0                     
#> 3 1         mock da… cohort_1          cohort_1           0                     
#> 4 5         mock da… cohort_2          cohort_3           0                     
#> 5 4         mock da… cohort_2          cohort_2           0                     
#> 6 6         mock da… cohort_3          cohort_1           0                     
#> 7 8         mock da… cohort_3          cohort_3           0                     
#> 8 7         mock da… cohort_3          cohort_2           0                     
#> # ℹ 15 more variables: washout_window <chr>, index_marker_gap <chr>,
#> #   combination_window <chr>, confidence_interval <chr>, restriction <chr>,
#> #   index_first_pharmac_count <chr>, index_first_pharmac_percentage <chr>,
#> #   marker_first_pharmac_count <chr>, marker_first_pharmac_percentage <chr>,
#> #   crude_sequence_ratio_point_estimate <chr>,
#> #   adjusted_sequence_ratio_point_estimate <chr>,
#> #   crude_sequence_ratio_lower_CI <chr>, crude_sequence_ratio_upper_CI <chr>, …
CDMConnector::cdmDisconnect(cdm = cdm)
# }