Intersecting the index and marker cohorts prior to calculating Sequence Symmetry Ratios

Join two tables in the CDM (one for index and the other for marker cohorts) into a new table in the cdm taking into account the maximum time interval between events. Index and marker cohorts should be instantiated in advance by the user.

Usage

generateSequenceCohortSet(
  cdm,
  indexTable,
  markerTable,
  name,
  cohortDateRange = as.Date(c(NA, NA)),
  indexId = NULL,
  markerId = NULL,
  daysPriorObservation = 0,
  washoutWindow = 0,
  indexMarkerGap = NULL,
  combinationWindow = c(0, 365)
)

Arguments

cdm

A CDM reference.

indexTable

A table in the CDM that the index cohorts should come from.

markerTable

A table in the CDM that the marker cohorts should come from.

name

The name within the cdm that the output is called. Default is joined_cohorts.

cohortDateRange

Two dates indicating study period and the sequences that the user wants to restrict to.

indexId

Cohort definition IDs in indexTable to be considered for the analysis. Change to NULL if all indices are wished to be included.

markerId

Cohort definition IDs in markerTable to be considered for the analysis. Change to NULL if all markers are wished to be included.

daysPriorObservation

The minimum amount of prior observation required on both the index and marker cohorts per person.

washoutWindow

A washout window to be applied on both the index cohort event and marker cohort.

indexMarkerGap

The maximum allowable gap between the end of the first episode and the start of the second episode in a sequence/combination.

combinationWindow

A constrain to be placed on the gap between two initiations. Default c(0,365), meaning the gap should be larger than 0 but less than or equal to 365.

Value

A table within the cdm reference.

Examples

# \donttest{
library(CohortSymmetry)
cdm <- mockCohortSymmetry()
#> Note: method with signature ‘DBIConnection#Id’ chosen for function ‘dbExistsTable’,
#>  target signature ‘duckdb_connection#Id’.
#>  "duckdb_connection#ANY" would also be valid
cdm <- generateSequenceCohortSet(
  cdm = cdm,
  name = "joined_cohorts",
  indexTable = "cohort_1",
  markerTable = "cohort_2"
)
 cdm$joined_cohorts
#> # Source:   table<main.joined_cohorts> [?? x 6]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1024-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date index_date
#>                   <int>      <dbl> <date>            <date>          <date>    
#>  1                    5          2 2022-05-22        2022-05-31      2022-05-22
#>  2                    6          2 2022-05-22        2022-05-25      2022-05-22
#>  3                    9          5 2019-04-07        2020-02-29      2019-04-07
#>  4                    3          4 2021-06-01        2022-05-25      2021-06-01
#>  5                    2          4 2021-05-25        2021-06-01      2021-06-01
#>  6                    7          1 2020-12-30        2021-01-01      2021-01-01
#>  7                    1          3 2009-09-09        2010-01-01      2009-09-09
#>  8                    2          5 2019-08-01        2020-05-25      2019-08-01
#>  9                    3          5 2019-08-01        2020-02-29      2019-08-01
#> 10                    2          1 2020-04-01        2021-01-01      2020-04-01
#> # ℹ more rows
#> # ℹ 1 more variable: marker_date <date>
 CDMConnector::cdmDisconnect(cdm = cdm)
# }