A formatted visualization of sequence_symmetry objects.

It provides a formatted table with the contents of the summariseSequenceRatio output.

Usage

tableSequenceRatios(
  result,
  type = "gt",
  estimateNameFormat = c(`N (%)` = "<count> (<percentage> %)", `SR (CI)` =
    "<point_estimate> (<lower_CI> - <upper_CI>)"),
  style = "default",
  crude = TRUE,
  adjusted = TRUE,
  studyPopulation = TRUE,
  indexName = TRUE,
  markerName = TRUE,
  cdmName = TRUE,
  .options = NULL
)

Arguments

result

A sequence_symmetry object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

estimateNameFormat

The columns that the user wishes to see for the formatted table, by default it would display both the counts and sequence ratios.

style

Named list that specifies how to style the different parts of a gt table or flextable. See visOmopResults package for more information on how to define a style. Alternatively, use "default" to get visOmopResults style, or NULL for gt/flextable default styling.

crude

whether to report crude results.

adjusted

whether to report adjusted results.

studyPopulation

whether to report the study population.

indexName

whether to report index names.

markerName

whether to report marker names

cdmName

whether to report database names.

.options

named list with additional formatting options. tableSequenceRatiosOptions() shows allowed arguments and their default values.

Value

A formatted version of the sequence_symmetry object.

Examples

# \donttest{
library(CohortSymmetry)
cdm <- mockCohortSymmetry()
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      89% | ETA:  0s
cdm <- generateSequenceCohortSet(cdm = cdm,
                                                 indexTable = "cohort_1",
                                                 markerTable = "cohort_2",
                                                 name = "joined_cohort")
res <- summariseSequenceRatio(cohort = cdm$joined_cohort)
#> Warning: For at least some combinations, index is always before marker or marker always
#> before index
#> -- 5 combinations of 8 had index always before marker
#> -- 5 combinations of 8 had marker always before index
#> Joining with `by = join_by(days_prior_observation, washout_window,
#> index_marker_gap, combination_window, confidence_interval,
#> moving_average_restriction, cdm_name)`
#> ! The following column type were changed:
#> • result_id: from character to integer
gtResult <- tableSequenceRatios(res)
CDMConnector::cdmDisconnect(cdm = cdm)
# }